refactor(permission): PermissionCommandRepository 이름 변경 - PermissionCommandRepository를 도메인 영역으로 옮기면서 PermissionRepository로 이름을 바꿨습니다. - PermissionCommandRepository를 사용하던 클래스에서 PermissionRepository로 변경했습니다.
