+++ src/main/java/com/simpleboard/board/permission/application/command/PermissionCommandService.java
@@ -1,7 +1,7 @@
 package com.simpleboard.board.permission.application.command;
 
 import com.simpleboard.board.permission.application.command.dto.DelegateRoleCommand;
-import com.simpleboard.board.permission.application.command.repository.PermissionCommandRepository;
+import com.simpleboard.board.permission.domain.PermissionRepository;
 import com.simpleboard.board.permission.application.common.UserFetchService;
 import com.simpleboard.board.permission.domain.PermissionPolicy;
 import org.springframework.stereotype.Service;
@@ -14,30 +14,30 @@ import org.springframework.stereotype.Service;
  */
 @Service
 public class PermissionCommandService {
-  private final PermissionCommandRepository permissionCommandRepository;
+  private final PermissionRepository permissionRepository;
   private final UserFetchService userFetchService;
 
   public PermissionCommandService(
-      PermissionCommandRepository permissionCommandRepository, UserFetchService userFetchService) {
-    this.permissionCommandRepository = permissionCommandRepository;
+      PermissionRepository permissionRepository, UserFetchService userFetchService) {
+    this.permissionRepository = permissionRepository;
     this.userFetchService = userFetchService;
   }
 
   public void delegateRole(Long boardId, DelegateRoleCommand delegateRoleCommand) {
-    PermissionPolicy permissionPolicy = permissionCommandRepository.getByBoardId(boardId);
+    PermissionPolicy permissionPolicy = permissionRepository.getByBoardId(boardId);
     Long toMemberId = userFetchService.getUserIdByNickname(delegateRoleCommand.toUserNickname());
     permissionPolicy.delegateRole(
         delegateRoleCommand.fromUserId(), toMemberId, delegateRoleCommand.roleName());
-    permissionCommandRepository.save(permissionPolicy);
+    permissionRepository.save(permissionPolicy);
   }
 
   public void deletePermissionPolicy(Long boardId) {
-    PermissionPolicy permissionPolicy = permissionCommandRepository.getByBoardId(boardId);
-    permissionCommandRepository.delete(permissionPolicy);
+    PermissionPolicy permissionPolicy = permissionRepository.getByBoardId(boardId);
+    permissionRepository.delete(permissionPolicy);
   }
 
   public void createPermissionPolicy(Long boardId, Long userID) {
     PermissionPolicy permissionPolicy = PermissionPolicy.create(boardId, userID);
-    permissionCommandRepository.save(permissionPolicy);
+    permissionRepository.save(permissionPolicy);
   }
 }
